requirements.txt
flask==2.3.3
flask-cors==4.0.0
reportlab==4.0.4
openpyxl==3.1.2
selenium==4.15.2
webdriver-manager==4.0.2
urllib3==2.0.7
trio==0.22.0
trio-websocket==0.11.1
rpaframework-core==11.3.3
robotframework==6.1.1
robotframework-seleniumlibrary==6.1.0
pywin32==306; sys_platform == "win32"